Rename ot.py to optimal_transport.py
