+++ fmralign/methods/__init__.py
@@ -1,5 +1,5 @@
 from .identity import Identity
-from .ot import OptimalTransport, SparseUOT
+from .optimal_transport import OptimalTransport, SparseUOT
 from .procrustes import Procrustes
 from .ridge import RidgeAlignment
 from .srm import DetSRM
