+++ fmralign/methods/tests/test_ot.py
@@ -2,7 +2,7 @@ import numpy as np
 import torch
 from numpy.testing import assert_array_almost_equal
 
-from fmralign.methods.ot import (
+from fmralign.methods.optimal_transport import (
     OptimalTransport,
     SparseUOT,
 )
